+++ wcmatch/glob.py
@@ -113,11 +113,6 @@ class Glob(object):
         if not self.pattern and self.npattern:
             self.pattern.append(_wcparse.WcPathSplit((b'**' if self.is_bytes else '**'), self.flags).split())
 
-    def _is_globstar(self, name):
-        """Check if name is a rescursive globstar."""
-
-        return self.globstar and name in (b'**', '**')
-
     def _is_hidden(self, name):
         """Check if is file hidden."""
 
@@ -186,14 +181,14 @@ class Glob(object):
                         try:
                             if (not dir_only or f.is_dir()) and matcher(f.name):
                                 yield os.path.join(curdir, f.name)
-                        except OSError:
+                        except OSError:  # pragma: no cover
                             pass
             else:
                 for f in os.listdir(scandir):
                     is_dir = os.path.isdir(os.path.join(curdir, f))
                     if (not dir_only or is_dir) and matcher(f):
                         yield os.path.join(curdir, f)
-        except OSError:
+        except OSError:  # pragma: no cover
             pass
 
     def _glob_deep(self, curdir, matcher, dir_only=False):
@@ -223,7 +218,7 @@ class Glob(object):
                                 yield path
                             if f.is_dir():
                                 yield from self._glob_deep(path, matcher, dir_only)
-                        except OSError:
+                        except OSError:  # pragma: no cover
                             pass
             else:
                 for f in os.listdir(scandir):
@@ -236,7 +231,7 @@ class Glob(object):
                         yield path
                     if is_dir:
                         yield from self._glob_deep(path, matcher, dir_only)
-        except OSError:
+        except OSError:  # pragma: no cover
             pass
 
     def _glob(self, curdir, this, rest):
@@ -264,17 +259,20 @@ class Glob(object):
             # Throw away multiple consecutive globstars
             # and acquire the pattern after the globstars if available.
             this = rest.pop(0) if rest else None
-            done = this is None
-            if done:
-                target = None
-            while this and not done:
+            globstar_end = this is None
+            while this and not globstar_end:
                 if this:
                     dir_only = this.dir_only
                     target = this.pattern
                 if this and this.is_globstar:
                     this = rest.pop(0) if rest else None
+                    if this is None:
+                        globstar_end = True
                 else:
-                    done = True
+                    break
+
+            if globstar_end:
+                target = None
 
             # We match `**/next` during `_glob_deep`, so what ever comes back,
             # we will send back through `_glob` with pattern after `next` (`**/next/after`).
@@ -286,11 +284,13 @@ class Glob(object):
             # Deep glob will account for this.
             matcher = self._get_matcher(target)
 
-            # If our pattern ends with `**` it matches zero or more,
+            # If our pattern ends with `curdir/**`, but does not start with `**` it matches zero or more,
             # so it should return `curdir/`, signifying curdir + no match.
             # If a pattern follows `**/something`, we always get the appropriate
             # return already, so this isn't needed in that case.
-            if matcher is None and curdir:
+            # There is one quirk though with Bash, if `curdir` had magic before `**`, Bash
+            # omits the trailing `/`. We don't worry about that.
+            if globstar_end and curdir:
                 yield os.path.join(curdir, self.empty)
 
             # Search
@@ -365,8 +365,9 @@ class Glob(object):
                     # on a case sensitive file system may return more than
                     # one starting location.
                     results = [curdir] if this.is_drive else self._get_starting_paths(curdir)
-                    if not results:
+                    if not results:  # pragma: no cover
                         # Nothing to do.
+                        # Do we need this?
                         return
 
                     if this.dir_only: