+++ docs/src/markdown/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.0.1
+
+- **FIX**: Ensure that all patterns in `glob` that have a directory preceding `**` but also end with `**` returns the preceding directory.
+- **FIX**: Fix byte conversion in path normalization.
+- **FIX**: Ensure POSIX character classes, when at the start of a sequence, properly have hyphens escaped following it. `[[:ascii:]-z]` should convert to `[\x00-\x7f\\-b]` not `[\x00-\x7f-b]`.
+- **FIX**: Fix an issue where we would fail because we couldn't covert raw characters even though raw character parsing was disabled.
+- **FIX**: Better default for file patterns.  Before if no pattern was provided for files, `'*'` was assumed, now it is `''`, and if `''` is used, all files will be matched. This works better for when full path is enabled as you get the same file matching logic.
+
 ## 1.0.0
 
 - Initial release